+++ tests/testthat/test_approx_near_optimal_survey_scheme.R
@@ -169,7 +169,7 @@ test_that("multiple species (sparse)", {
 
 test_that("consistent results", {
   skip_if_not_installed("RandomFields")
-  skip_on_os("Windows")
+  skip_on_os("windows")
   # seeds
   set.seed(505)
   RandomFields::RFoptions(seed = 505)
@@ -218,7 +218,7 @@ test_that("consistent results", {
 test_that("consistent results (multiple threads)", {
   # skip if using PSOCK cluster and package not installed
   skip_if_not_installed("RandomFields")
-  skip_on_os("Windows")
+  skip_on_os("windows")
   skip_if(!requireNamespace("surveyvoi") &&
           !identical(.Platform$OS.type, "unix"))
   # seeds