+++ tests/testthat/test_evdsi.R
@@ -2,7 +2,7 @@ context("evdsi")
 
 test_that("single species", {
   skip_if_not_installed("RandomFields")
-  skip_on_os("Windows")
+  skip_on_os("windows")
   # data
   ## set seeds
   set.seed(123)
@@ -63,7 +63,7 @@ test_that("single species", {
 
 test_that("multiple species", {
   skip_if_not_installed("RandomFields")
-  skip_on_os("Windows")
+  skip_on_os("windows")
   # data
   ## set seeds
   set.seed(123)