+++ tests/testthat/test_optimal_survey_scheme.R
@@ -3,7 +3,7 @@ context("optimal_survey_scheme")
 test_that("single species", {
   skip_if_not_installed("gurobi")
   skip_if_not_installed("RandomFields")
-  skip_on_os("Windows")
+  skip_on_os("windows")
   # data
   set.seed(500)
   site_data <- sf::st_as_sf(
@@ -175,7 +175,7 @@ test_that("multiple species (sparse)", {
 test_that("consistent results", {
   skip_if_not_installed("gurobi")
   skip_if_not_installed("RandomFields")
-  skip_on_os("Windows")
+  skip_on_os("windows")
   # seeds
   set.seed(505)
   RandomFields::RFoptions(seed = 505)
@@ -224,7 +224,7 @@ test_that("consistent results", {
 test_that("consistent results (multiple threads)", {
   skip_if_not_installed("gurobi")
   skip_if_not_installed("RandomFields")
-  skip_on_os("Windows")
+  skip_on_os("windows")
   # skip if using PSOCK cluster and package not installed
   skip_if(!requireNamespace("surveyvoi") &&
           !identical(.Platform$OS.type, "unix"))