+++ tests/testthat/test_rcpp_approx_expected_value_of_decision_given_survey_scheme.R
@@ -2,7 +2,7 @@ context("rcpp_approx_expected_value_of_decision_given_survey_scheme")
 
 test_that("single species", {
   skip_if_not_installed("RandomFields")
-  skip_on_os("Windows")
+  skip_on_os("windows")
   # data
   ## set seeds
   set.seed(123)
@@ -66,7 +66,7 @@ test_that("single species", {
 
 test_that("multiple species", {
   skip_if_not_installed("RandomFields")
-  skip_on_os("Windows")
+  skip_on_os("windows")
   # data
   ## set seeds
   set.seed(123)